feat(lua): assay.vault stdlib client for assay-engine vault
Plan 17 §S11. Pure-Lua wrapper at crates/assay/stdlib/vault.lua that
talks to the engine's /api/v1/vault/* HTTP surface. Mirrors the plan-
locked surface for this phase: KV v2 lifecycle (put/get/list/meta/
delete/destroy/undelete) and transit (create/encrypt/decrypt/rotate/
list).

Defaults: engine_url + admin_key fall back to ASSAY_ENGINE_URL +
ASSAY_ADMIN_API_KEY. Encrypt/decrypt take + return raw Lua strings
(binary-safe via base64 on the wire) so callers don't have to know
about the envelope encoding.

Naming: plan 17 took `vault` as the assay-native module name. The
existing HashiCorp Vault / OpenBao client (which targets the upstream
HC Vault HTTP API) is renamed to `assay.hashicorp_vault`; the
existing `assay.openbao` alias and all existing tests follow.
Per-conversation: "no backward compatibility — active development".

Tests:
- crates/assay/tests/stdlib_vault.rs — 7 wiremock-backed cases
  covering KV put/get round-trip, version-specific GET, lifecycle
  paths (delete/destroy/undelete query-string version), list with
  and without prefix, transit create/encrypt/decrypt/rotate, missing
  engine_url surfaces a Lua error, 401 propagates as a Lua error
  with the HTTP status in the message.
- crates/assay/tests/stdlib_hashicorp_vault.rs — renamed from
  stdlib_vault.rs; all 50 require() calls updated.
- crates/assay/tests/{fs_require,stdlib_postgres,stdlib_openbao_alias}
  + crates/assay/tests/e2e/v050_features.lua — sed-updated to point
  at assay.hashicorp_vault.

All renamed + new Lua tests green.
