+++ crates/assay/stdlib/openbao.lua
@@ -3,5 +3,5 @@
 --- @keywords openbao, vault, secrets, kv, policies, auth, transit, pki, encryption, decryption, certificate, seal, initialization, authentication, secret-engine, password, rotation
 
 -- OpenBao alias: OpenBao is API-compatible with HashiCorp Vault.
--- Both tools can use the same client via require("assay.vault") or require("assay.openbao").
-return require("assay.vault")
+-- Both tools can use the same client via require("assay.hashicorp_vault") or require("assay.openbao").
+return require("assay.hashicorp_vault")
